project2/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
future==0.18.2
idna==3.2
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7
